% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.scitation.R
\name{as.scitation}
\alias{as.scitation}
\title{coerce things to scitations}
\usage{
as.scitation(x)
}
\arguments{
\item{x}{input}
}
\description{
coerce things to scitations
}
\examples{
\dontrun{
# with rcrossref
library("rcrossref")
res <- cr_works(limit = 10)
res$data$type <- "article"
as.scitation(res$data) 

# self
x <- scitation("article", 'petkun2016', doi = "10.7717/peerj.1126")
as.scitation(x)

# a doi
as.scitation("10.7717/peerj.1126")
}
}

