% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_df.R
\name{as_df}
\alias{as_df}
\title{coerce scitations to data.frame}
\usage{
as_df(x)
}
\arguments{
\item{x}{input}
}
\description{
coerce scitations to data.frame
}
\examples{
#x <- scitation("article", 'petkun2016', doi = "10.7717/peerj.2399")
#as_df(x)

x <- scitation(
  "Article", "cui2016", 
  title = "Short and long term representation of an unfamiliar tone distribution",
  author = "Anja X. Cui and Charlette Diercks and Nikolaus F. Troje and Lola L. Cuddy", 
  journaltitle = "The Journal Title",
  date = "2016-08-30")
as_df(x)

# from DOI
# as_df(as.scitation("10.7717/peerj.1126"))
}

