% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sciprint.R
\name{sciprint}
\alias{sciprint}
\title{print citation}
\usage{
sciprint(x)
}
\arguments{
\item{x}{input}
}
\description{
print citation
}
\examples{
sciprint(scitation("article", 'petkun2016', doi = "10.7717/peerj.1126"))
}

