% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scitation.R
\name{scitation}
\alias{scitation}
\title{Make a scitation}
\usage{
scitation(type, key = NULL, ..., .list = list())
}
\arguments{
\item{type}{type of work}

\item{key}{a key for your entry}

\item{...}{key-value pairs, depends on work type, see 
\code{\link{work_types}}}
}
\description{
Make a scitation
}
\examples{
# scitation("article", doi = "10.7717/peerj.1126")
scitation(
 "article", 
 'petkun2016', 
 doi = "10.7717/peerj.1126",
 author = "foobar",
 title = "That and this and stuff and things",
 journaltitle = "Stuff And Things",
 year = 2001
)

scitation(
 "book", 
 'deGroot2015', 
 doi = "10.1007/978-1-349-22928-4",
 author = "de Groot, Joanna; Maynard, Mary",
 title = "Women's Studies in the 1990s",
 year = 2015
)
}
