% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scitation_df.R
\name{scitation_df}
\alias{scitation_df}
\title{Coerce scitations to a data.frame}
\usage{
scitation_df(..., .list = NULL)
}
\arguments{
\item{...}{objects of class \code{scitation}}

\item{.list}{a list, with only objects of class \code{scitation}}
}
\description{
Coerce scitations to a data.frame
}
\examples{
#x <- scitation("article", 'petkun2016', doi = "10.7717/peerj.2399")
#scitation_df(x)

x <- scitation(
  "Article", "cui2016", 
  title = "Short and long term representation of an unfamiliar tone distribution",
  author = "Anja X. Cui and Charlette Diercks and Nikolaus F. Troje and Lola L. Cuddy", 
  journaltitle = "The Journal Title",
  date = "2016-08-30")
  
# any number of inputs of class scitation
scitation_df(x, x, x)

# inputs of class scitation in a list
scitation_df(.list = list(x, x, x, x))

# from DOI
# scitation_df(as.scitation("10.7717/peerj.1126"))
}

