% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scitations-package.r
\docType{package}
\name{scitations-package}
\alias{scitations}
\alias{scitations-package}
\title{Citations for R.}
\description{
Citations for R.
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}

