% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scitation.R
\name{scitation}
\alias{scitation}
\title{Make a scitation}
\usage{
scitation(type, key = NULL, ...)
}
\arguments{
\item{type}{type of work}

\item{key}{a key for your entry}

\item{...}{key-value pairs, depends on work type}
}
\description{
Make a scitation
}
\examples{
scitation("article", doi = "10.7717/peerj.1126")
scitation("article", 'petkun2016', doi = "10.7717/peerj.1126")
}

