% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scitations-package.r
\docType{package}
\name{scitations-package}
\alias{scitations}
\alias{scitations-package}
\title{Citations for R.}
\description{
\pkg{scitations} is meant to make it easy to work with citations, whether 
in tidy format or not, and to convert from various formats to citations, 
and to various formats.
}
\details{
The package introduces a simple S3 class called a `scitation` that we 
can then perform many operations with or on.

There are other bibliographic packages out there - this package tries 
not to duplicate their efforts, while striving for a fun to use
citation workflow.
}
\section{Scitations}{

The package is called \pkg{scitations}, but we also refer to the citation
objects by the same name. The objects are simple S3 classes, for which 
you can access any element with \code{$}, or \code{[}, or \code{[[}.
}

\section{Package API}{

\itemize{
 \item scitation - create a scitation
 \item as.scitation - coerce objects to a scitation
 \item scitation_df - make a set of tidy scitations
 \item work_types - get work types metadata, and required fields
}
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}

